Create README.md
Electrocardiogram(ECG) is a method of measuring the electrical activities of heart. Every portion of ECG is very essential for the diagnosis of different cardiac problems. But the amplitude and duration of ECG Signal is usually corrupted by different noises. In this project, we will deal with denoising the corrupted ECG signal at hand. The Normalized-Least-Mean-Squares algorithm based adaptive filter is used for this purpose. In many applications involving noise cancellation, the changes in signal characteristics are quite fast. This requires the utilization of adaptive algorithms which converge rapidly. Normalized Least Mean Squares adaptive filters have been used in a wide range of signal processing application because of its simplicity in computation and implementation. The NLMS algorithm has an advantage of low computational complexity. There is always a trade-off between computational complexity and fast convergence when it comes to adaptive filtering algorithms. We approach the NLMS algorithm by updating its filter co-efficient at each time instant, to fulfill a suitable trade-off between convergence rate and computational complexity. The performance of the proposed NLMS algorithm is fully studied through the analysis used in adaptive filters and general expressions.